+++ CHANGELOG.md
@@ -16,6 +16,15 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Breaking
 
+## 2025-07-01
+
+- yellowstone-grpc-geyser-8.1.0
+
+### Features
+
+- geyser plugin exposes new metrics to measure subscriber performance such send/recv bandwidth load.
+- Added metrics to measure the amount of account update and sizes we are receiving from agave.
+
 ## 2025-06-30
 
 - @triton-one/yellowstone-grpc@4.1.0